Create README.md
This project is a part of Mall Customer Segmentation Data competition held on Kaggle. The goal of K means is to group data points into distinct non-overlapping subgroups. One of the major application of K means clustering is segmentation of customers to get a better understanding of them which in turn could be used to increase the revenue of the company.